Corrected file structure
